strategy pattern refactor
