+++ Behavioral/Strategy/index.html
@@ -1,23 +1,1 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Strategy Pattern</title>
-    <script src="./dist/strategyPattern.js" defer></script>
-    <link rel="stylesheet" href="style.css" />
-  </head>
-  <body>
-    <div class="window">
-      <nav></nav>
-      <h3>Payment Method</h3>
-      <div class="payment-list">
-        <div id="card" class="payment-button">Credit Card</div>
-        <div id="paypal" class="payment-button">Paypal</div>
-        <div id="bank" class="payment-button">Bank Account</div>
-      </div>
-      <h3>Total Cost: <span class="total"> 0</span></h3>
-    </div>
-  </body>
-</html>
+<!doctype html><html lang="en"><head><meta charset="UTF-8"/><meta http-equiv="X-UA-Compatible" content="IE=edge"/><meta name="viewport" content="width=device-width,initial-scale=1"/><title>Strategy Pattern</title><link rel="stylesheet" href="style.css"/><script defer="defer" src="dist/index.js"></script></head><body><div class="window"><nav></nav><h3>Payment Method</h3><div class="payment-list"><div id="card" class="payment-button">Credit Card</div><div id="paypal" class="payment-button">Paypal</div><div id="bank" class="payment-button">Bank Account</div></div><h3>Total Cost: <span class="total">0</span></h3></div></body></html>
